# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BioGene Unified Bioinformatics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lims/samples:
    get:
      operationId: listSamples
      tags: [lims]
      summary: List all biological samples
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Samples list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SampleListResponse"
    post:
      operationId: createSample
      tags: [lims]
      summary: Create a new biological sample
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSampleBody"
      responses:
        "201":
          description: Created sample
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
  /lims/samples/{sampleId}:
    get:
      operationId: getSample
      tags: [lims]
      summary: Get a single sample
      parameters:
        - name: sampleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sample detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
    patch:
      operationId: updateSample
      tags: [lims]
      summary: Update a sample
      parameters:
        - name: sampleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSampleBody"
      responses:
        "200":
          description: Updated sample
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
    delete:
      operationId: deleteSample
      tags: [lims]
      summary: Delete a sample
      parameters:
        - name: sampleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /lims/experiments:
    get:
      operationId: listExperiments
      tags: [lims]
      summary: List all experiments
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Experiments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentListResponse"
    post:
      operationId: createExperiment
      tags: [lims]
      summary: Create a new experiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperimentBody"
      responses:
        "201":
          description: Created experiment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
  /lims/experiments/{experimentId}:
    get:
      operationId: getExperiment
      tags: [lims]
      summary: Get a single experiment
      parameters:
        - name: experimentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Experiment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
    patch:
      operationId: updateExperiment
      tags: [lims]
      summary: Update an experiment
      parameters:
        - name: experimentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExperimentBody"
      responses:
        "200":
          description: Updated experiment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
...
components:
  schemas:
...
    Sample:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        concentration:
          type: number
        unit:
          type: string
        volume:
          type: number
        organism:
          type: string
        tissue:
          type: string
        storageLocation:
          type: string
        barcode:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, status, createdAt]
    SampleListResponse:
      type: object
      properties:
        samples:
          type: array
          items:
            $ref: "#/components/schemas/Sample"
        total:
          type: integer
      required: [samples, total]
    CreateSampleBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        concentration:
          type: number
        unit:
          type: string
        volume:
          type: number
        organism:
          type: string
        tissue:
          type: string
        storageLocation:
          type: string
        notes:
          type: string
      required: [name, type]
    UpdateSampleBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        concentration:
          type: number
        volume:
          type: number
        storageLocation:
          type: string
        notes:
          type: string
    Experiment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        sampleIds:
          type: array
          items:
            type: string
        protocol:
          type: string
        notes:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, status, createdAt]
    ExperimentListResponse:
      type: object
      properties:
        experiments:
          type: array
          items:
            $ref: "#/components/schemas/Experiment"
        total:
          type: integer
      required: [experiments, total]
    CreateExperimentBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        sampleIds:
          type: array
          items:
            type: string
        protocol:
          type: string
        notes:
          type: string
      required: [name, type]
    UpdateExperimentBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        protocol:
          type: string
        notes:
          type: string
        completedAt:
          type: string
...